chore: change game implementation
The previous game.js was committed but not used. Here, I use the Game
class to implement a more general game mechanics and options.
The previous RPSGame class contained the game mechanics. Instead it
contains the specific mechanics for RPS. It is close to the previous
RPSInstance class which is gone now.
I refactor the app.js file to reduce amount of code.
However, for now, the IA implementation is not yet done.